backend/pyproject.toml
[tool.poetry]
name = "backend"
version = "0.1.0"
description = ""
authors = ["Miguel de Freitas Muniz <miguimuniz@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "3.12.*"
fastapi = {extras = ["standard"], version = "^0.115.0"}
sqlalchemy = "^2.0.35"
pydantic-settings = "^2.5.2"
psycopg = {extras = ["binary"], version = "^3.2.3"}
alembic = "^1.13.3"


[tool.poetry.group.dev.dependencies]
pytest = "^8.3.3"
pytest-cov = "^5.0.0"
taskipy = "^1.13.0"
ruff = "^0.6.9"

[tool.ruff]
line-length = 79
extend-exclude = ['migrations']

[tool.ruff.lint]
preview = true
select = ['I', 'F', 'E', 'W', 'PL', 'PT']

[tool.ruff.format]
preview = true
quote-style = 'single'

[tool.pytest.ini_options]
pythonpath = "."
addopts = '-p no:warnings'

[tool.taskipy.tasks]
lint = 'ruff check . && ruff check . --diff'
format = 'ruff check . --fix && ruff format .'
run = 'fastapi dev backend/app.py'
pre_test = 'task lint'
test = 'pytest -s -x --cov=backend -vv'
post_test = 'coverage html'
revision = "alembic revision --autogenerate"
upgrade = "alembic upgrade head"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
